Initial commit of bootstrap script
This script should enable us to install ModLE on any machine with:
 - A recent version of Python3 (probably something like python 3.5 or
   newer)
 - A C++17 compiler toolchain (e.g. GCC 8.3 or Clang 9.0)

For the time being the script only builds ModLE, but does not install
it, nor it cleans temporary files. This will be fixed at some point in
the future.
The idea is that wants to compile ModLE will only have to download the
source code and run utils/bootstrap.sh.
This script will then:
 - Create a Python venv
 - Install Conan
 - Check whether a recent version of CMake is available, and if it
 isn't, use Conan to build a recent version of CMake
 - Download, configure and install ModLE's dependencies. This is all
   done by Conan while configuring the build using CMake
- Compile modle and modle_tools using -DBUILD_SHARED_LIBS=OFF
- Install the binaries and remove all temporary files