Fix bug that would cause modle sim to hang when ncells < nthreads
This bug was caused by attempting to post n tasks to a task queue with
capacity < n.
In the previous implementation getting the queue capacity right was
crucial: if the queue was too small, some worker threads would have to
wait for tasks to become available, while with an excessively large
queue we would waste a lot of memory.
The memory waste was not due to the internal buffer used by the queue,
but rather by the fact that we were allocating the contact matrix for
a given chromosome when creating the first task for said chromosome.

In the new implementation, contact matrices are allocated by the first
worker thread that executes a task for a given chromosome.
We are using a mutex to ensure that if multiple threads attempt to
concurrently (de)allocate a contact matrix, only one succeeds.

Given the above, we now have a bit more freedom when deciding the task
queue size, which means that we don't need to handle edge cases where
the queue size is >> than the ncells or nthreads.