Update burn-in code to leverage the CLI args introduced with the pev. commit
The changes in this commit are trying to address the following issue:
Running MoDLE twice on the same data using the target contact density as
stopping criterion, while keeping all the parameters unchanged except
for --ncells should produce nearly identical results (modulo the small
differences due to stochastic effects).
This turns out not to be the case, as contact matrices produced with a
low number of cells (e.g. 32) have a tendency to accumulate slightly
more contacts further away from the diagonal when compared to contact
matrices produced with a higher number of cells (e.g. 5000).

This is most likely due to the fact that when a small number of cells
are being simulated, each simulation instance will last longer (i.e.
will run for more epochs) than when the same simulation is carried out
on more cells, as in the first case it will take longer for each simulation
instance to reach the target contact density.

Before this commit, with all settings @default and --ncells=5000, the
burn-in phase for a single cell would last ~240 epochs, while the actual
simulation would only take ~10 epochs.
Instead, when --ncells=32, MoDLE will spend ~240 epochs in burn-in phase
and ~1500 in the actual simulation.
Assuming an overall avg. extr speed of 10kbp/epoch, the upper-bound for
loop length is ~2.4Mbp for --ncells=5000, while it is ~15Mbp when
--ncells=32. While @default settings we don't expect any loop to reach a
size of 15Mbp, we do expect a small fraction of loops to have a size >
2.4Mbp. That being said, when --ncells=5000 we won't be able to observe
said loops, and this is the underlying issue producing the unexpected
contact matrix differences we are observing.

A simple solution to this problem is to have longer burn-ins, especially
when --ncells is a large number. Unfortunately this would likely
significantly increase simulation runtimes.

Changes introduced by this commit:
--burn-in-extr-speed-coefficient can be used to increase the extrusion
speed during the burn-in phase.
When --burn-in-lef-activation-epochs and --burn-in-epochs are left at
default, MoDLE will compute the target number of burn-in epochs based on
the avg. LEF lifetime and extrusion speed(s). The number of burn-in
epochs is larger for higher avg. LEF lifetimes and/or slower extr.
speeds.
The effective extrusion speed used during the burn-in phase is
calculated by multiplying the rev/fwd extr. speed by the coefficient
specified through --burn-in-extr-speed-coefficient, meaning that when
this coefficient is > 1.0 loop extrusion will be faster during the
burn-in phase. This should (and does) not affect the time required
to simulate one burn-in epoch.

When increasing extr. speed, we are making a trade-off between speed and
accuracy, which should be fine in the case of the burn-in phase.
In case the accuracy loss is unacceptable for some of the use cases,
users can pass --burn-in-extr-speed-coefficient=1 to turn off this
optimization.