Rework how LEFs are released
This commit changes the way LEFs are selected and released.
The main reason behind this change was to address the problems observed
while simulating loop extrusion using a small number of LEFs (e.g. 200
LEFs on chromosome 1). Under the above conditions we observed an
unexpected and significant concentration of contacts around the
diagonal.

Before this commit, every epoch we computed the average number of LEFs
to be released as (nlefs * extr. speed) / lef-lifetime.
This number was then used as the lambda parameter of a Poisson PRNG to
obtain the actual number of LEFs to be released during the current
epoch.
The actual LEFs to be released are selected using a weighted-uniform
distribution (i.e. std::discrete_distribution).

The Poiss. distribution worked well enough for large number of LEFs, not
so much for a relatively low number of LEFs, like 200.

After this commit, LEFs will be released according to the following
logic.
We introduce the parameter --prob-of-lef-release. This defines the
"base" probability that a LEF will be released during a given epoch.
Every epoch MoDLE will loop over the active LEFs and run a Bernoulli
trial with probability of success equal to --prob-of-lef-release.
Whenever a trial succeds, the corresponding LEF is released.
The probability of success is affected by the LEF state.
For example, if a LEF in involved in a hard-stall and
--hard-stall-multiplier=5 and --prob-of-lef-release=0.015, then the
probability of success will be equal to 0.015 * (1.0 / 5.0).