Bugfix
Fixed a bug in MoDLE simulation code that caused severe artifacts in the
output matrix when simulating loop extrusion with a low LEF density.
The bug was always present regardless of LEF density but it was much
less noticeable with normal/high LEF densities.

Changes from this commit are causing one test unit to fail. I have
already identified the failure reason and am working on a fix.

[no ci]