Implement LockRange* class
LockRangeExclusive and LockRangeShared implement locking of a range of
mutexes in contiguous memory using RAII to avoid deadlocks.
Upon construction, both classes acquire all mutexes provided as param to
the constructor. Mutexes are always acquired from begin() to end().
Mutexes are unlocked in the destructor.
If an exception is thrown in the constructor all mutexes acquired so far
by the constructor are released before re-throwing the exception.