Remove unused code and minor updates to ContactMatrix class
- Remove code that was used by modle_tools stats
- Use std::mutex instead of std::shared_mutex
- To lock the entire matrix acquire all the mutexes in the _mtxes vector
  instead of locking the global_mutex
- Add test case for the global locking

NOTE: TSAN sometimes reports a data race in ContactMatrix::increment()
when running MoDLE. It looks like a false positive to me.
I also have not been able to reproduce the race condition in a test
case.
MoDLE's output does not seem to be affected by the data race, further
supporting the hypothesis that this is a false positive (as a data race
in the increment method should lead to slightly different contact
matrices)