Replace --prob-of-lef-release with --avg-lef-processivity
The main motivation for this change is to improve usability.
Changing extrusion speed without also changing --prob-of-lef-release
had the (potentially) unintended consequence of changing LEF
processivity.
The probability of LEF release Pr is now computed with:
Pr = s / Proc

where s is the bidirectional extrusion speed, and Proc is the average
LEF processivity