Rework project versioning one more time
Project version is defined in cmake/Versioning.cmake
If the project looks like a git repository, then we use
cmake-git-version-tracking to collect git status metadata.
This metadata can then be included in the version shown by
modle --version and modle_tools --version.

If the project doesn't contain a .git folder, then we assume the project
is being compiled without modifications, and thus only version
information specified in cmake/Versioning.cmake is used.

There are two main differences between code from this commit, and code
from e17b704:
- In the latest commit, most logic has been moved to C++ code inside
  src/config/
- To avoid recompiling all files including either version.hpp or
  git.hpp every time the git repository metadata changes, the actual
  metadata is now embedded in .cpp files, and is access through several
  accessor functions. These functions are either trivial (i.e. consist
  of just a return statement), or use std::call_once and static
  variables to avoid recomputing the same value at every function call

The latest versioning strategy is less flexible than the the previous
one (as git metadata is embedded in the binary and is not available to CMake),
but is less buggy and has negligible effects on compilation times.