Bugfix
Address a bug that caused the rev/fwd extrusion speed to not be updated
based on the bin size passed through the --resolution option.

I decided to update the integration test script to use 2.5kbp as average
extrusion speed instead of updating the test dataset.
The main motivation is to limit repository bloat due to many revision of
(relatively) large binary files.