Rewrite extrusion barrier class
Extrusion barriers are now represented using two classes:

- ExtrusionBarrier: the "old" class. Used to represent one extrusion
  barrier instance.
- ExtrusionBarriers: the "new" class. Used to represent extrusion
  barriers for a given chromosome.

ExtrusionBarriers was written with the aim of improving locality of
reference by using a struct of vectors instead of a vector of structs to
represent extrusion barriers on a chromosome.
The extrusion barrier state (active/not active) is now stored in the
ExtrusionBarriers class instead of in a separate bitset.

Changes introduced by this commit should yield identical output to code
from acb60ea.
I verified this is the case with default settings and with uniform
extrusion barrier occupancy.