Normalize transition/collision probabilities base on avg extrusion speed
In brief, before this commit, changing the average extrusion speed had
non-obvious effects on the output matrices.
In general, simulating at lower extrusion speeds resulted in weaker
stripes and dots, as well as higher contact dispersion.
An examplanation of why this was the case, is provided in the help
message for --probability-normalization-factor.

This commit introduces the --probability-normalization-factor CLI
option, which is used to normalize collision and transition
probabilities.