Rework contact sampling logic
Rework contact sampling to deal with very small simulations (e.g. those
involving a single LEF).

With a single or few LEFs, it is possible (and even likely), that none
of the LEFs is within bound when sampling interactions. In this case,
the contact sampling loop would run indefinitely.
Changes from this commit addresst this issue.

IMPORTANT: changes from this commit changes MoDLE's output!